u-boot-script: add a new option for USB boot support on Meson8*
To enable it needs to set a variable in the uboot environment:
'fw_setenv enableusbboot 1' (default: 0)

This option is not enabled by default, because scanning USB ports in
some cases increases the boot time with additional delay. And it is
important to note that some devices cannot boot from USB at all, or
only certain USB ports can be used for this purpose.

In addition, it is also possible to disable the SD card boot option:
'fw_setenv enablesdboot 0' (default: 1)